Update on "Consistent compute numel/contiguous strategy with SymInts"
Previously, our handling for contiguity was inconsistent in the following ways:

- is_strides_like 2d/3d and is_non_overlapping_and_dense always were computed
  based on sizes_and_strides_, even if you had symbolic ints
- Furthermore, even if you set custom policy for strides, these quantities were
  not overridable by subclasses
- Furthermore, we didn't even store these fields on ExtraMeta
- We duplicate implementations of compute_contiguous (plain, channels last,
  channels last 3d)
- We inconsistently called refresh_numel()/refresh_contiguous(), versus
  recomputing it ourselves

This factor makes a consistent strategy for all of the boolean fields, and
for numel computation.  After this refactor:

- All layout boolean fields are interposable via strides policy
  and can be overridden from Python; you will never access a garbage field
- All layout boolean fields are on ExtraMeta
- You can always call refresh_numel/contiguous, no matter if your Tensor is
  contiguous or not
- The numel/layout boolean fields are always populated consistently with
  the sizes strides fields (either on Tensor or ExtraMeta), even if you
  have custom policy
- There is only one implementation of the actual computation logic


Differential Revision: [D39907696](https://our.internmc.facebook.com/intern/diff/D39907696)

[ghstack-poisoned]

+++ c10/core/TensorImpl.cpp
@@ -236,7 +236,8 @@ bool_is_contiguous _compute_contiguous(
   if (numel == 0)
     return bool_is_contiguous(is_contiguous);
   T z = 1;
-  for (int64_t d = sizes.size() - 1; d >= 0; d--) {
+  // NB: make sure we do signed arithmetic
+  for (int64_t d = int64_t(sizes.size()) - 1; d >= 0; d--) {
     const auto size_d = sizes[d];
     if (size_d != 1) {
       if (strides[d] == z) {
